+++ 1.21.11/src/main/kotlin/com/bleudev/nine_lifes/api/Types.kt
@@ -1,4 +1,0 @@
-package com.bleudev.nine_lifes.api
-
-typealias Transformer<T> = (T) -> T
-typealias EmptyPredicate = () -> Boolean