+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # 5.2
+
+## Version changes
+- End of support for 1.21.11 (bleudev [#102](https://github.com/bleudev/nine_lifes/pull/102))
+
 ## New features
-- Charge 5 amethyst shards with lightning at a time (bleudev [#99](https://github.com/bleudev/nine_lifes/pull/99))
+- Charge `max_charged_items_at_a_time` amethyst shards with lightning at a time (bleudev [#99](https://github.com/bleudev/nine_lifes/pull/99))
+- Add 5 new gamerules (bleudev [#102](https://github.com/bleudev/nine_lifes/pull/102)):
+  - `take_lifes`
+  - `take_lifes_in_overworld`
+  - `take_lifes_in_nether`
+  - `take_lifes_in_end`
+  - `max_charged_items_at_a_time`
 
 ## Fixes
 - Delete some debug prints (bleudev [#99](https://github.com/bleudev/nine_lifes/pull/99))